Fixed check_burst bug
Now we keep track of the oldest timestamp in every window helper.

If check_burst is set to True, we do not generate another correlation alert untill the window is expired.

In WeakWindowHelper and StrongWindowHelper this is done with the help of the variable _oldestTimestamp
that tracks the timestamp of the last correlation alert.