Fixed severe bug in idmef lack
When a lot of events are received, the null idmef will be never reached
it is necessary to introduce a new check that if time_now - start_consider_timestamp < 1 return
otherwise process the idmef anyway, but set idmef to None. In this case when we receive something that is not
in our filters, we consider that idmef as None, and process idmef as None, but the processing is done anyway.